% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkBars.R
\name{BarCloses}
\alias{BarCloses}
\title{Gets the bars' closed as requested}
\usage{
BarCloses(barsVar)
}
\arguments{
\item{barsVar}{RHost variable that stores bar array}
}
\description{
Gets the bars' closed as requested
}

