% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkBars.R
\name{BarHighs}
\alias{BarHighs}
\title{Gets the bars' high  as requested}
\usage{
BarHighs(barsVar)
}
\arguments{
\item{barsVar}{RHost variable that stores bar array}
}
\description{
Gets the bars' high  as requested
}

