% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkRLib.R
\name{BarLows}
\alias{BarLows}
\title{Gets the bars' low as requested}
\usage{
BarLows(barsVar)
}
\arguments{
\item{barsVar}{RHost variable that stores bar array}
}
\description{
Gets the bars' low as requested
}

