% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkBars.R
\name{BarOpens}
\alias{BarOpens}
\title{Gets the bars' open as requested}
\usage{
BarOpens(barsVar)
}
\arguments{
\item{barsVar}{RHost variable that stores bar array}
}
\description{
Gets the bars' open as requested
}

