% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkRLib.R
\name{BarTos}
\alias{BarTos}
\title{Gets the bars' volume as requested}
\usage{
BarTos(barsVar)
}
\arguments{
\item{barsVar}{RHost variable that stores bar array}
}
\description{
Gets the bars' volume as requested
}

