% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkBars.R
\name{ComputeBars}
\alias{ComputeBars}
\title{Gets the bars as requested range}
\usage{
ComputeBars(symbol, priceTypeStr, barPeriodStr, startTime, barCountDbl)
}
\arguments{
\item{symbol}{Symbol looked}

\item{priceTypeStr}{Ask}

\item{barPeriodStr}{Values like: M1, H1}

\item{endTimeEpoch}{Epoch time}
}
\description{
Gets the bars as requested range
}

