% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkBars.R
\name{ComputeBarsRange}
\alias{ComputeBarsRange}
\title{Gets the bars as requested}
\usage{
ComputeBarsRange(symbol, priceTypeStr, barPeriodStr, startTime, endTime,
  barCount)
}
\arguments{
\item{symbol}{Symbol looked}

\item{priceTypeStr}{Ask}

\item{barPeriodStr}{Values like: M1, H1}

\item{startTime}{Start of the time range}

\item{endTime}{End of the time range}

\item{barCount}{Items used}
}
\description{
Gets the bars as requested
}

