% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkBarPair.R
\name{ComputeGetPairBars}
\alias{ComputeGetPairBars}
\title{Gets the bars pairs as requested}
\usage{
ComputeGetPairBars(symbol, barPeriodStr, startTime, barCountDbl)
}
\arguments{
\item{symbol}{Symbol looked}

\item{barPeriodStr}{Values like: M1, H1}

\item{barCountDbl}{Bar count}

\item{startTimeEpoch}{Epoch time}
}
\description{
Gets the bars pairs as requested
}

