% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkRLib.R
\name{ComputeGetPairBarsRange}
\alias{ComputeGetPairBarsRange}
\title{Gets the bars pairs as requested}
\usage{
ComputeGetPairBarsRange(symbol, barPeriodStr, startTimeEpoch, endTimeEpoch)
}
\arguments{
\item{symbol}{Symbol looked}

\item{barPeriodStr}{Values like: M1, H1}

\item{startTimeEpoch}{Epoch time}

\item{endTimeEpoch}{Epoch time}
}
\description{
Gets the bars pairs as requested
}

