% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkRLib.R
\name{ComputeQuoteHistory}
\alias{ComputeQuoteHistory}
\title{Gets the bars as requested}
\usage{
ComputeQuoteHistory(symbol, startTimeEpoch, endTimeEpoch, depth)
}
\arguments{
\item{symbol}{Symbol looked}

\item{startTimeEpoch}{Starting time. Use ttGetEpochFromText if you want to take from text a valid date.}

\item{endTimeEpoch}{Ending time. Use ttGetEpochFromText if you want to take from text a valid date.}

\item{depth}{Quotes depth}
}
\description{
Gets the bars as requested
}

