% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkBarPair.R
\name{getBarPairFrame}
\alias{getBarPairFrame}
\title{Extracts bar pair array data as a full data frame}
\usage{
getBarPairFrame(bars)
}
\arguments{
\item{bars}{Bars array variable}
}
\description{
Extracts bar pair array data as a full data frame
}

