% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkBarPair.R
\name{GetBarsAskClose}
\alias{GetBarsAskClose}
\title{Gets the bars' ask as requested}
\usage{
GetBarsAskClose(barsPairVar)
}
\arguments{
\item{quotesVar}{RHost variable that stores quotes array}
}
\description{
Gets the bars' ask as requested
}

