% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkBarPair.R
\name{GetBarsAskLow}
\alias{GetBarsAskLow}
\title{Gets the bars' ask as requested}
\usage{
GetBarsAskLow(barsPairVar)
}
\arguments{
\item{quotesVar}{RHost variable that stores quotes array}
}
\description{
Gets the bars' ask as requested
}

