% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkRLib.R
\name{GetBarsBidHigh}
\alias{GetBarsBidHigh}
\title{Gets the bars' ask as requested}
\usage{
GetBarsBidHigh(barsPairVar)
}
\arguments{
\item{quotesVar}{RHost variable that stores quotes array}
}
\description{
Gets the bars' ask as requested
}

