% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkRLib.R
\name{GetBarsBidOpen}
\alias{GetBarsBidOpen}
\title{Gets the bars' ask as requested}
\usage{
GetBarsBidOpen(barsPairVar)
}
\arguments{
\item{quotesVar}{RHost variable that stores quotes array}
}
\description{
Gets the bars' ask as requested
}

