% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkBars.R
\name{getBarsFrame}
\alias{getBarsFrame}
\title{Extracts bar array data as a full data frame}
\usage{
getBarsFrame(symbolBars)
}
\arguments{
\item{symbolBars}{Bars array variable}
}
\description{
Extracts bar array data as a full data frame
}

