% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkTrades.R
\name{GetTradeAgentCommission}
\alias{GetTradeAgentCommission}
\title{Get trade comission}
\usage{
GetTradeAgentCommission(symInfo)

GetTradeAgentCommission(symInfo)

GetTradeAgentCommission(symInfo)
}
\description{
Get trade comission

Get trade comission

Get trade comission
}

