% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkQuotes.r
\name{QuotesCreatingTime}
\alias{QuotesCreatingTime}
\title{Gets the bars' ask as requested}
\usage{
QuotesCreatingTime(quotesVar)
}
\arguments{
\item{quotesVar}{RHost variable that stores quotes array}
}
\description{
Gets the bars' ask as requested
}

