% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkRLib.R
\name{QuotesHasAsk}
\alias{QuotesHasAsk}
\title{Gets the quotes' ask as requested}
\usage{
QuotesHasAsk(quotesVar)
}
\arguments{
\item{quotesVar}{RHost variable that stores quotes array}
}
\description{
Gets the quotes' ask as requested
}

