% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkQuotes.r
\name{QuotesHasBid}
\alias{QuotesHasBid}
\title{Gets the quotes' bid as requested}
\usage{
QuotesHasBid(quotesVar)
}
\arguments{
\item{quotesVar}{RHost variable that stores quotes array}
}
\description{
Gets the quotes' bid as requested
}

