% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkQuotes.r
\name{QuotesSpread}
\alias{QuotesSpread}
\title{Gets the quotes' spread as requested}
\usage{
QuotesSpread(quotesVar)
}
\arguments{
\item{quotesVar}{RHost variable that stores quotes array}
}
\description{
Gets the quotes' spread as requested
}

