% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkBarPair.R
\name{ttBarPairsByCount}
\alias{ttBarPairsByCount}
\title{Gets the bars' low as requested}
\usage{
ttBarPairsByCount(symbol, barPeriodStr, startTime, barCountDbl = 10000)
}
\arguments{
\item{symbol}{Symbol looked}

\item{barPeriodStr}{Values like: M1, H1}

\item{startTimeEpoch}{Epoch time}

\item{barCount}{Bar count}
}
\description{
Gets the bars' low as requested
}

