% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkRLib.R
\name{ttConnect}
\alias{ttConnect}
\title{Connects to a TT server}
\usage{
ttConnect(address = "", login = "", password = "")
}
\arguments{
\item{address}{Url of the running server}

\item{login}{Account number you login}

\item{password}{Password for the account you login}
}
\description{
Connects to a TT server
}

