% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkConnect.R
\name{ttDisconnect}
\alias{ttDisconnect}
\title{Disconnect from a TT server}
\usage{
ttDisconnect()
}
\arguments{
\item{address}{Url of the running server}

\item{login}{Account number you login}

\item{password}{Password for the account you login}
}
\description{
Disconnect from a TT server
}

