% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkConnect.R
\name{ttDisplayDate}
\alias{ttDisplayDate}
\title{Displays a DateTime}
\usage{
ttDisplayDate(dateToShow)
}
\arguments{
\item{dateToShow}{An R datetime}
}
\description{
Displays a DateTime
}

