% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkBarPair.R
\name{ttGetBarPairsRange}
\alias{ttGetBarPairsRange}
\title{Gets the bars' low as requested}
\usage{
ttGetBarPairsRange(symbol, barPeriodStr, startTimeEpoch, endTimeEpoch)
}
\arguments{
\item{symbol}{Symbol looked}

\item{barPeriodStr}{Values like: M1, H1}

\item{startTimeEpoch}{Epoch time}

\item{endTimeEpoch}{Epoch time}
}
\description{
Gets the bars' low as requested
}

