% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkBars.R
\name{ttGetBars}
\alias{ttGetBars}
\title{Gets the bars' low as requested}
\usage{
ttGetBars(symbol, priceTypeStr, barPeriodStr, endTime = 0, barCount = 1e+06)

ttGetBars(symbol, priceTypeStr, barPeriodStr, endTime = 0, barCount = 1e+06)
}
\arguments{
\item{symbol}{Symbol looked ('EURUSD')}

\item{priceTypeStr}{Ask, Bid}

\item{barPeriodStr}{Values like: M1, M5, H1}

\item{barCount}{Count of bars. Default: 10000000}

\item{mode}{"range" is default. Mode "list" allows just to get a range of items}

\item{symbol}{Symbol looked}

\item{priceTypeStr}{Ask}

\item{barPeriodStr}{Values like: M1, H1}
}
\description{
Gets the bars' low as requested

Gets the bars' low as requested
}

