% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkRLib.R
\name{ttGetBarsInfo}
\alias{ttGetBarsInfo}
\title{Gets the bars pairs as requested}
\usage{
ttGetBarsInfo(symbol, priceTypeStr, barPeriodStr)
}
\arguments{
\item{symbol}{Symbol looked}

\item{priceTypeStr}{Ask or bid}

\item{barPeriodStr}{Values like: M1, H1}
}
\description{
Gets the bars pairs as requested
}

