% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkRLib.R
\name{ttGetEpochFromText}
\alias{ttGetEpochFromText}
\title{Gets the epoch (double) time from a .Net date}
\usage{
ttGetEpochFromText(currentTime)
}
\arguments{
\item{currentTime}{.Net invariant time}
}
\description{
Gets the epoch (double) time from a .Net date
}

