% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkConnect.R
\name{ttInit}
\alias{ttInit}
\title{Initialize the CLR runtime and loads the FDK host assembly}
\usage{
ttInit()
}
\description{
Initialize the CLR runtime and loads the FDK host assembly
}

