% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkQuotes.r
\name{ttSetBarCount}
\alias{ttSetBarCount}
\title{Sets the bar count inside calls}
\usage{
ttSetBarCount(barCount)
}
\arguments{
\item{barCount}{Set a different bar count than the default (1 million)}
}
\description{
Sets the bar count inside calls
}

