% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkRLib.R
\name{ttUnregisterAllVariables}
\alias{ttUnregisterAllVariables}
\title{Clear the environment variables}
\usage{
ttUnregisterAllVariables(varName)
}
\arguments{
\item{varName}{.Net variable to be removed}
}
\description{
Clear the environment variables
}

