% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkRLib.R
\name{ComputeBars}
\alias{ComputeBars}
\title{Gets the bars as requested}
\usage{
ComputeBars(symbol, priceTypeStr, barPeriodStr)
}
\arguments{
\item{symbol}{Symbol looked}

\item{priceTypeStr}{Ask}

\item{barPeriodStr}{Values like: M1, H1}
}
\description{
Gets the bars as requested
}

