% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkRLib.R
\name{ComputeGetPairBars}
\alias{ComputeGetPairBars}
\title{Gets the bars pairs as requested}
\usage{
ComputeGetPairBars(symbol, barPeriodStr, startTimeEpoch)
}
\arguments{
\item{symbol}{Symbol looked}

\item{barPeriodStr}{Values like: M1, H1}

\item{startTimeEpoch}{Epoch time}
}
\description{
Gets the bars pairs as requested
}

