% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkBars.R
\name{ttBarsByCount}
\alias{ttBarsByCount}
\title{Gets the bars' low as requested}
\usage{
ttBarsByCount(symbol, priceTypeStr = "Bid", barPeriodStr = "M1", startTime,
  barCount = 10000)
}
\arguments{
\item{symbol}{Symbol looked}

\item{priceTypeStr}{Ask or Bid}

\item{barPeriodStr}{Values like: M1, H1}

\item{startTime}{Time to start}

\item{barCount}{(defaults to 10 000)}
}
\description{
Gets the bars' low as requested
}

