% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkRLib.R
\name{ttGetBars}
\alias{ttGetBars}
\title{Gets the bars' low as requested}
\usage{
ttGetBars(symbol, priceTypeStr, barPeriodStr)

ttGetBars(symbol, priceTypeStr, barPeriodStr)
}
\arguments{
\item{symbol}{Symbol looked}

\item{priceTypeStr}{Ask}

\item{barPeriodStr}{Values like: M1, H1}

\item{symbol}{Symbol looked}

\item{priceTypeStr}{Ask}

\item{barPeriodStr}{Values like: M1, H1}
}
\description{
Gets the bars' low as requested

Gets the bars' low as requested
}

