% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FdkRLib.R
\name{ttSetBarCount}
\alias{ttSetBarCount}
\title{Gets the bars' ask as requested}
\usage{
ttSetBarCount(barCount)
}
\arguments{
\item{barCount}{Set a different bar count than the default (1 million)}
}
\description{
Gets the bars' ask as requested
}

